[#4306] test(e2e): extract the shared auth-journey helpers from the redirect spec
Pure extraction: submitSignup, verifyInFreshContext, signIn,
waitForAppReady, waitForPathname and FRESH_CONTEXT move to
e2e/support/auth-journey.ts unchanged so the plan-intent journey can
reuse them. No behavior change.
